+++ CHANGELOG.md
@@ -1,14 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.7.1] -2018-05-31
+## [0.8.0] -2018-05-31
 
 #### Changed
+* ProductModel now extends Isotope\Model\Product\Standard
 * removed heimrichhannot/contao-request dependency
 * removed some unnecassary services
 * dependency injection for download helper
 * enhanced code quality for some classes
 * updated some namespaces
+
+#### Fixed
+* booking plan reservation not working
 
 ## [0.7.0] -2018-05-31